EnvStats Functions for Calibration

The EnvStats functions listed below are useful for performing calibration and inverse prediction to determine the concentration of a chemical based on a machine signal.

Details

Function Name	Description

anovaPE	Compute lack-of-fit and pure error ANOVA table for a

linear model.

calibrate	Fit a calibration line or curve.

detectionLimitCalibrate	Determine detection limit based on using a calibration

line (or curve) and inverse regression.

inversePredictCalibrate	Predict concentration using a calibration line (or curve)

and inverse regression.

pointwise	Pointwise confidence limits for predictions.

predict.lm	Predict method for linear model fits.