EnvStats Functions for Censored Data

The EnvStats functions listed below are useful for dealing with Type I censored data.

Details

Data Transformations

Function Name	Description

boxcoxCensored	Compute values of an objective for Box-Cox Power

transformations, or compute optimal transformation,

for Type I censored data.

print.boxcoxCensored	Print an object of class "boxcoxCensored".

plot.boxcoxCensored	Plot an object of class "boxcoxCensored".

Estimating Distribution Parameters

Function Name	Description

egammaCensored	Estimate shape and scale parameters for a gamma distribution

based on Type I censored data.

egammaAltCensored	Estimate mean and CV for a gamma distribution

based on Type I censored data.

elnormCensored	Estimate parameters for a lognormal distribution (log-scale)

based on Type I censored data.

elnormAltCensored	Estimate parameters for a lognormal distribution (original scale)

based on Type I censored data.

enormCensored	Estimate parameters for a Normal distribution based on Type I

censored data.

epoisCensored	Estimate parameter for a Poisson distribution based on Type I

censored data.

enparCensored	Estimate the mean and standard deviation nonparametrically.

gpqCiNormSinglyCensored	Generate the generalized pivotal quantity used to construct a

confidence interval for the mean of a Normal distribution based

on Type I singly censored data.

gpqCiNormMultiplyCensored	Generate the generalized pivotal quantity used to construct a

confidence interval for the mean of a Normal distribution based

on Type I multiply censored data.

print.estimateCensored	Print an object of class "estimateCensored".

Estimating Distribution Quantiles

Function Name	Description

eqlnormCensored	Estimate quantiles of a Lognormal distribution (log-scale)

based on Type I censored data, and optionally construct

a confidence interval for a quantile.

eqnormCensored	Estimate quantiles of a Normal distribution

based on Type I censored data, and optionally construct

a confidence interval for a quantile.

All of the functions for computing quantiles (and associated confidence intervals) for complete (uncensored) data are listed in the help file Estimating Distribution Quantiles. All of these functions, with the exception of eqnpar, will accept an object of class "estimateCensored". Thus, you may estimate quantiles (and construct approximate confidence intervals) for any distribution for which:

There exists a function to estimate distribution parameters using censored data (see the section Estimating Distribution Parameters above).

There exists a function to estimate quantiles for that distribution based on complete data (see the help file Estimating Distribution Quantiles).

Nonparametric estimates of quantiles (and associated confidence intervals) can be constructed from censored data as long as the order statistics used in the results are above all left-censored observations or below all right-censored observations. See the help file for eqnpar for more information and examples.

Goodness-of-Fit Tests

Function Name	Description

gofTestCensored	Perform a goodness-of-fit test based on Type I left- or

right-censored data.

print.gofCensored	Print an object of class "gofCensored".

plot.gofCensored	Plot an object of class "gofCensored".

Hypothesis Tests

Function Name	Description

twoSampleLinearRankTestCensored	Perform two-sample linear rank tests based on

censored data.

print.htestCensored	Printing method for object of class

"htestCensored".

Plotting Probability Distributions

Function Name	Description

cdfCompareCensored	Plot two cumulative distribution functions based on Type I

censored data.

ecdfPlotCensored	Plot an empirical cumulative distribution function based on

Type I censored data.

ppointsCensored	Compute plotting positions for Type I censored data.

qqPlotCensored	Produce quantile-quantile (Q-Q) plots, also called probability

plots, based on Type I censored data.

Prediction and Tolerance Intervals

Function Name	Description

gpqTolIntNormSinglyCensored	Generate the generalized pivotal quantity used to construct a

tolerance interval for a Normal distribution based

on Type I singly censored data.

gpqTolIntNormMultiplyCensored	Generate the generalized pivotal quantity used to construct a

tolerance interval for a Normal distribution based

on Type I multiply censored data.

tolIntLnormCensored	Tolerance interval for a lognormal distribution (log-scale)

based on Type I censored data.

tolIntNormCensored	Tolerance interval for a Normal distribution based on Type I

censored data.

All of the functions for computing prediction and tolerance intervals for complete (uncensored) data are listed in the help files Prediction Intervals and Tolerance Intervals. All of these functions, with the exceptions of predIntNpar and tolIntNpar, will accept an object of class "estimateCensored". Thus, you may construct approximate prediction or tolerance intervals for any distribution for which:

There exists a function to estimate distribution parameters using censored data (see the section Estimating Distribution Parameters above).

There exists a function to create a prediction or tolerance interval for that distribution based on complete data (see the help files Prediction Intervals and Tolerance Intervals).

Nonparametric prediction and tolerance intervals can be constructed from censored data as long as the order statistics used in the results are above all left-censored observations or below all right-censored observations. See the help files for predIntNpar, predIntNparSimultaneous, and tolIntNpar for more information and examples.