EnvStats Functions for Goodness-of-Fit Tests

The EnvStats functions listed below are useful for performing goodness-of-fit tests for user-specified probability distributions.

Details

Goodness-of-Fit Tests

Function Name	Description

gofTest	Perform a goodness-of-fit test for a specified probability distribution.

The resulting object is of class "gof" unless the test is the

two-sample Kolmogorov-Smirnov test, in which case the resulting

object is of class "gofTwoSample".

plot.gof	S3 class method for plotting an object of class "gof".

print.gof	S3 class method for printing an object of class "gof".

plot.gofTwoSample	S3 class method for plotting an object of class "gofTwoSample".

print.gofTwoSample	S3 class method for printing an object of class "gofTwoSample".

gofGroupTest	Perform a goodness-of-fit test to determine whether data in a set of groups

appear to all come from the same probability distribution

(with possibly different parameters for each group).

The resulting object is of class "gofGroup".

plot.gofGroup	S3 class method for plotting an object of class "gofGroup".

print.gofGroup	S3 class method for printing an object of class "gofGroup".

Tests for Outliers

Function Name	Description

rosnerTest	Perform Rosner's test for outliers assuming a normal (Gaussian) distribution.

print.gofOutlier	S3 class method for printing an object of class "gofOutlier".

Choose a Distribution

Function Name	Description

distChoose	Choose best fitting distribution based on goodness-of-fit tests.

print.distChoose	S3 class method for printing an object of class "distChoose".