EnvStats Functions for Hypothesis Tests

The EnvStats functions listed below are useful for performing hypothesis tests not already built into R. See Power and Sample Size Calculations for a list of functions you can use to perform power and sample size calculations based on various hypothesis tests.

Details

For goodness-of-fit tests, see Goodness-of-Fit Tests.

Function Name	Description

chenTTest	Chen's modified one-sided t-test for skewed

distributions.

kendallTrendTest	Nonparametric test for monotonic trend

based on Kendall's tau statistic (and

optional confidence interval for slope).

kendallSeasonalTrendTest	Nonparametric test for monotonic trend

within each season based on Kendall's tau

statistic (and optional confidence interval

for slope).

oneSamplePermutationTest	Fisher's one-sample randomization

(permutation) test for location.

quantileTest	Two-sample rank test to detect a shift in

a proportion of the “treated” population.

quantileTestPValue	Compute p-value associated with a specified

combination of \(m\), \(n\), \(r\) and \(k\)

for the quantile test.

Useful for determining \(r\) and \(k\) for a

given significance level \(\alpha\).

serialCorrelationTest	Test for the presence of serial correlation.

signTest	One- or paired-sample sign test on the

median.

twoSampleLinearRankTest	Two-sample linear rank test to detect a

shift in the “treated” population.

twoSamplePermutationTestLocation	Two-sample or paired-sample randomization

(permutation) test for location.

twoSamplePermutationTestProportion	Randomization (permutation) test to compare

two proportions (Fisher's exact test).

varTest	One-sample test on variance or two-sample

test to compare variances.

varGroupTest	Test for homogeneity of variance among two

or more groups.

zTestGevdShape	Estimate the shape parameter of a

Generalized Extreme Value distribution and

test the null hypothesis that the true

value is equal to 0.