EnvStats Functions for Prediction Intervals

The EnvStats functions listed below are useful for computing prediction intervals and simultaneous prediction intervals. See Power and Sample Size for a list of functions useful for computing power and sample size for a design based on a prediction interval width, or a design based on a hypothesis test for future observations falling outside of a prediciton interval.

Details

Function Name	Description

predIntGamma,	Prediction interval for the next \(k\)

predIntGammaAlt	observations or next set of \(k\) means for a

Gamma distribution.

predIntGammaSimultaneous,	Construct a simultaneous prediction interval for the

predIntGammaAltSimultaneous	next \(r\) sampling occasions based on a

Gamma distribution.

predIntLnorm,	Prediction interval for the next \(k\)

predIntLnormAlt	observations or geometric means from a

Lognormal distribution.

predIntLnormSimultaneous,	Construct a simultaneous prediction interval for the

predIntLnormAltSimultaneous	next \(r\) sampling occasions based on a

Lognormal distribution.

predIntNorm	Prediction interval for the next \(k\) observations

or means from a Normal (Gaussian) distribution.

predIntNormK	Compute the value of \(K\) for a prediction interval

for a Normal distribution.

predIntNormSimultaneous	Construct a simultaneous prediction interval for the

next \(r\) sampling occasions based on a

Normal distribution.

predIntNormSimultaneousK	Compute the value of \(K\) for a simultaneous

prediction interval for the next \(r\) sampling

occasions based on a Normal distribution.

predIntNpar	Nonparametric prediction interval for the next \(k\)

of \(K\) observations.

predIntNparSimultaneous	Construct a nonparametric simultaneous prediction

interval for the next \(r\) sampling occasions.

predIntPois	Prediction interval for the next \(k\) observations

or sums from a Poisson distribution.