EnvStats Functions for Printing and Plotting Objects of Various S3 Classes

The EnvStats functions listed below are printing and plotting methods for various S3 classes.

Details

Printing Methods

Function Name	Description

print.boxcox	Print an object that inherits from class "boxcox".

print.boxcoxCensored	Print an object that inherits from class

"boxcoxCensored".

print.boxcoxLm	Print an object that inherits from class "boxcoxLm".

print.estimate	Print an object that inherits from class "estimate".

print.estimateCensored	Print an object that inherits from class

"estimateCensored".

print.gof	Print an object that inherits from class "gof".

print.gofCensored	Print an object that inherits from class "gofCensored".

print.gofGroup	Print an object that inherits from class "gofGroup".

print.gofTwoSample	Print an object that inherits from class

"gofTwoSample".

print.htest	Print an object that inherits from class "htest".

print.htestCensored	Print an object that inherits from class

"htestCensored".

print.permutationTest	Print an object that inherits from class

"permutationTest".

print.summaryStats	Print an object that inherits from class

"summaryStats".

Plotting Methods

Function Name	Description

plot.boxcox	Plot an object that inherits from class "boxcox".

plot.boxcoxCensored	Plot an object that inherits from class "boxcoxCensored".

plot.boxcoxLm	Plot an object that inherits from class "boxcoxLm".

plot.gof	Plot an object that inherits from class "gof".

plot.gofCensored	Plot an object that inherits from class "gofCensored".

plot.gofGroup	Plot an object that inherits from class "gofGroup".

plot.gofTwoSample	Plot an object that inherits from class "gofTwoSample".

plot.permutationTest	Plot an object that inherits from class "permutationTest".