EnvStats Probability Distributions and Random Numbers

Listed below are all of the probability distributions available in R and EnvStats. Distributions with a description in bold are new ones that are part of EnvStats. For each distribution, there are functions for generating: values for the probability density function, values for the cumulative distribution function, quantiles, and random numbers.

The data frame Distribution.df contains information about all of these probability distributions.

Details

Distribution Abbreviation	Description

beta	Beta distribution.

binom	Binomial distribution.

cauchy	Cauchy distribution.

chi	Chi distribution.

chisq	Chi-squared distribution.

exp	Exponential distribution.

evd	Extreme value distribution.

f	F-distribution.

gamma	Gamma distribution.

gammAlt	Gamma distribution parameterized with mean and CV.

gevd	Generalized extreme value distribution.

geom	Geometric distribution.

hyper	Hypergeometric distribution.

logis	Logistic distribution.

lnorm	Lognormal distribution.

lnormAlt	Lognormal distribution parameterized with mean and CV.

lnormMix	Mixture of two lognormal distributions.

lnormMixAlt	Mixture of two lognormal distributions

parameterized by their means and CVs.

lnorm3	Three-parameter lognormal distribution.

lnormTrunc	Truncated lognormal distribution.

lnormTruncAlt	Truncated lognormal distribution

parameterized by mean and CV.

nbinom	Negative binomial distribution.

norm	Normal distribution.

normMix	Mixture of two normal distributions.

normTrunc	Truncated normal distribution.

pareto	Pareto distribution.

pois	Poisson distribution.

t	Student's t-distribution.

tri	Triangular distribution.

unif	Uniform distribution.

weibull	Weibull distribution.

wilcox	Wilcoxon rank sum distribution.

zmlnorm	Zero-modified lognormal (delta) distribution.

zmlnormAlt	Zero-modified lognormal (delta) distribution

parameterized with mean and CV.

zmnorm	Zero-modified normal distribution.

In addition, the functions evNormOrdStats and evNormOrdStatsScalar compute expected values of order statistics from a standard normal distribution.