EnvStats Functions for Summary Statistics and Plots

The EnvStats functions listed below create summary statistics and plots.

Details

Summary Statistics
R comes with several functions for computing summary statistics, including mean, var, median, range, quantile, and summary. The following functions in EnvStats complement these R functions.

Function Name	Description

cv	Coefficient of variation

geoMean	Geometric mean

geoSD	Geometric standard deviation

iqr	Interquartile range

kurtosis	Kurtosis

lMoment	\(L\)-moments

pwMoment	Probability-weighted moments

skewness	Skew

summaryFull	Extensive summary statistics

summaryStats	Summary statistics

Summary Plots
R comes with several functions for creating plots to summarize data, including hist, barplot, boxplot, dotchart, stripchart, and numerous others.

The help file Plotting Probability Distributions lists several EnvStats functions useful for producing summary plots as well.

In addition, the EnvStats function stripChart is a modification of stripchart that allows you to include summary statistics on the plot itself.

Finally, the help file Plotting Using ggplot2 lists several EnvStats functions for adding information to plots produced with the ggplot function, including the function geom_stripchart, which is an adaptation of the EnvStats function stripChart.