The Generalized Extreme Value Distribution

Density, distribution function, quantile function, and random generation for the generalized extreme value distribution.

dgevd(x, location = 0, scale = 1, shape = 0)
  pgevd(q, location = 0, scale = 1, shape = 0)
  qgevd(p, location = 0, scale = 1, shape = 0)
  rgevd(n, location = 0, scale = 1, shape = 0)

Arguments

x

vector of quantiles.

q

vector of quantiles.

p

vector of probabilities between 0 and 1.

n

sample size. If length(n) is larger than 1, then length(n) random values are returned.

location

vector of location parameters.

scale

vector of positive scale parameters.

shape

vector of shape parameters.

Details

Let \(X\) be a generalized extreme value random variable with parameters location=\(\eta\), scale=\(\theta\), and shape=\(\kappa\). When the shape parameter \(\kappa = 0\), the generalized extreme value distribution reduces to the extreme value distribution. When the shape parameter \(\kappa \ne 0\), the cumulative distribution function of \(X\) is given by: $$F(x; \eta, \theta, \kappa) = exp\{-[1 - \kappa(x-\eta)/\theta]^{1/\kappa}\}$$ where \(-\infty < \eta, \kappa < \infty\) and \(\theta > 0\). When \(\kappa > 0\), the range of \(x\) is: $$-\infty < x \le \eta + \theta/\kappa$$ and when \(\kappa < 0\) the range of \(x\) is: $$\eta + \theta/\kappa \le x < \infty$$

The \(p^th\) quantile of \(X\) is given by: $$x_{p} = \eta + \frac{\theta \{1 - [-log(p)]^{\kappa}\}}{\kappa}$$

Value

density (devd), probability (pevd), quantile (qevd), or random sample (revd) for the generalized extreme value distribution with location parameter(s) determined by location, scale parameter(s) determined by scale, and shape parameter(s) determined by shape.

References

Forbes, C., M. Evans, N. Hastings, and B. Peacock. (2011). Statistical Distributions. Fourth Edition. John Wiley and Sons, Hoboken, NJ.

Jenkinson, A.F. (1955). The Frequency Distribution of the Annual Maximum (or Minimum) of Meteorological Events. Quarterly Journal of the Royal Meteorological Society, 81, 158–171.

Johnson, N. L., S. Kotz, and N. Balakrishnan. (1995). Continuous Univariate Distributions, Volume 2. Second Edition. John Wiley and Sons, New York.

Author

Steven P. Millard (EnvStats@ProbStatInfo.com)

Note

Two-parameter extreme value distributions (EVD) have been applied extensively since the 1930's to several fields of study, including the distributions of hydrological and meteorological variables, human lifetimes, and strength of materials. The three-parameter generalized extreme value distribution (GEVD) was introduced by Jenkinson (1955) to model annual maximum and minimum values of meteorological events. Since then, it has been used extensively in the hydological and meteorological fields.

The three families of EVDs are all special kinds of GEVDs. When the shape parameter \(\kappa = 0\), the GEVD reduces to the Type I extreme value (Gumbel) distribution. (The function zTestGevdShape allows you to test the null hypothesis that the shape parameter is equal to 0.) When \(\kappa > 0\), the GEVD is the same as the Type II extreme value distribution, and when \(\kappa < 0\) it is the same as the Type III extreme value distribution.

See also

Examples

  # Density of a generalized extreme value distribution with 
  # location=0, scale=1, and shape=0, evaluated at 0.5: 

  dgevd(.5) 
#> [1] 0.3307043
  #[1] 0.3307043

  #----------

  # The cdf of a generalized extreme value distribution with 
  # location=1, scale=2, and shape=0.25, evaluated at 0.5: 

  pgevd(.5, 1, 2, 0.25) 
#> [1] 0.2795905
  #[1] 0.2795905

  #----------

  # The 90'th percentile of a generalized extreme value distribution with 
  # location=-2, scale=0.5, and shape=-0.25: 

  qgevd(.9, -2, 0.5, -0.25) 
#> [1] -0.4895683
  #[1] -0.4895683

  #----------

  # Random sample of 4 observations from a generalized extreme value 
  # distribution with location=5, scale=2, and shape=1. 
  # (Note: the call to set.seed simply allows you to reproduce this example.)

  set.seed(20) 
  rgevd(4, 5, 2, 1) 
#> [1] 6.738692 6.473457 4.446649 5.727085
  #[1] 6.738692 6.473457 4.446649 5.727085