Ammonium Concentration in Precipitation Measured at Olympic National Park Hoh Ranger Station

Ammonium (NH\(_4\)) concentration (mg/L) in precipitation measured at Olympic National Park, Hoh Ranger Station (WA14), weekly or every other week from January 6, 2009 through December 20, 2011.

Olympic.NH4.df

Format

A data frame with 102 observations on the following 6 variables.

Date.On

Start of collection period. Date on which the sample bucket was installed on the collector.

Date.Off

End of collection period. Date on which the sample bucket was removed from the collector.

Week

a numeric vector indicating the cumulative week number starting from January 1, 2009.

NH4.Orig.mg.per.L

a character vector of the original NH\(_4\) concentrations reported either as the observed value or less than some detection limit. For values reported as less than a detection limit, the value reported is the actual limit of detection or, in the case of a diluted sample, the product of the detection limit value and the dilution factor.

NH4.mg.per.L

a numeric vector of NH\(_4\) concentrations with non-detects coded to their detection limit.

Censored

a logical vector indicating which observations are censored.

Details

StationOlympic National Park-Hoh Ranger Station (WA14)

LocationJefferson County, Washington

Latitude47.8597

Longitude-123.9325

Elevation182 meters

USGS 1:24000 Map NameOwl Mountain

Operating AgencyOlympic National Park

Sponsoring AgencyNPS-Air Resources Division

Source

National Atmospheric Deposition Program, National Trends Network (NADP/NTN).
https://nadp.slh.wisc.edu/data/sites/siteDetails.aspx?net=NTN&id=WA14