The Pareto Distribution

Density, distribution function, quantile function, and random generation for the Pareto distribution with parameters location and shape.

dpareto(x, location, shape = 1)
  ppareto(q, location, shape = 1)
  qpareto(p, location, shape = 1)
  rpareto(n, location, shape = 1)

Arguments

x

vector of quantiles.

q

vector of quantiles.

p

vector of probabilities between 0 and 1.

n

sample size. If length(n) is larger than 1, then length(n) random values are returned.

location

vector of (positive) location parameters.

shape

vector of (positive) shape parameters. The default is shape=1.

Details

Let \(X\) be a Pareto random variable with parameters location=\(\eta\) and shape=\(\theta\). The density function of \(X\) is given by: $$f(x; \eta, \theta) = \frac{\theta \eta^\theta}{x^{\theta + 1}}, \; \eta > 0, \; \theta > 0, \; x \ge \eta$$ The cumulative distribution function of \(X\) is given by: $$F(x; \eta, \theta) = 1 - (\frac{\eta}{x})^\theta$$ and the \(p\)'th quantile of \(X\) is given by: $$x_p = \eta (1 - p)^{-1/\theta}, \; 0 \le p \le 1$$ The mode, mean, median, variance, and coefficient of variation of \(X\) are given by: $$Mode(X) = \eta$$ $$E(X) = \frac{\theta \eta}{\theta - 1}, \; \theta > 1$$ $$Median(X) = x_{0.5} = 2^{1/\theta} \eta$$ $$Var(X) = \frac{\theta \eta^2}{(\theta - 1)^2 (\theta - 1)}, \; \theta > 2$$ $$CV(X) = [\theta (\theta - 2)]^{-1/2}, \; \theta > 2$$

Value

dpareto gives the density, ppareto gives the distribution function, qpareto gives the quantile function, and rpareto generates random deviates.

References

Forbes, C., M. Evans, N. Hastings, and B. Peacock. (2011). Statistical Distributions. Fourth Edition. John Wiley and Sons, Hoboken, NJ.

Johnson, N. L., S. Kotz, and N. Balakrishnan. (1994). Continuous Univariate Distributions, Volume 1. Second Edition. John Wiley and Sons, New York.

Author

Steven P. Millard (EnvStats@ProbStatInfo.com)

Note

The Pareto distribution is named after Vilfredo Pareto (1848-1923), a professor of economics. It is derived from Pareto's law, which states that the number of persons \(N\) having income \(\ge x\) is given by: $$N = A x^{-\theta}$$ where \(\theta\) denotes Pareto's constant and is the shape parameter for the probability distribution.

The Pareto distribution takes values on the positive real line. All values must be larger than the “location” parameter \(\eta\), which is really a threshold parameter. There are three kinds of Pareto distributions. The one described here is the Pareto distribution of the first kind. Stable Pareto distributions have \(0 < \theta < 2\). Note that the \(r\)'th moment only exists if \(r < \theta\).

The Pareto distribution is related to the exponential distribution and logistic distribution as follows. Let \(X\) denote a Pareto random variable with location=\(\eta\) and shape=\(\theta\). Then \(log(X/\eta)\) has an exponential distribution with parameter rate=\(\theta\), and \(-log\{ [(X/\eta)^\theta] - 1 \}\) has a logistic distribution with parameters location=\(0\) and scale=\(1\).

The Pareto distribution has a very long right-hand tail. It is often applied in the study of socioeconomic data, including the distribution of income, firm size, population, and stock price fluctuations.

See also

Examples

  # Density of a Pareto distribution with parameters location=1 and shape=1, 
  # evaluated at 2, 3 and 4: 

  dpareto(2:4, 1, 1) 
#> [1] 0.2500000 0.1111111 0.0625000
  #[1] 0.2500000 0.1111111 0.0625000

  #----------

  # The cdf of a Pareto distribution with parameters location=2 and shape=1, 
  # evaluated at 3, 4, and 5: 

  ppareto(3:5, 2, 1) 
#> [1] 0.3333333 0.5000000 0.6000000
  #[1] 0.3333333 0.5000000 0.6000000

  #----------

  # The 25'th percentile of a Pareto distribution with parameters 
  # location=1 and shape=1: 

  qpareto(0.25, 1, 1) 
#> [1] 1.333333
  #[1] 1.333333

  #----------

  # A random sample of 4 numbers from a Pareto distribution with parameters 
  # location=3 and shape=2. 
  # (Note: the call to set.seed simply allows you to reproduce this example.)

  set.seed(10) 
  rpareto(4, 3, 2)
#> [1] 4.274728 3.603148 3.962862 5.415322
  #[1] 4.274728 3.603148 3.962862 5.415322