S3 Class "boxcox"

Objects of S3 class "boxcox" are returned by the EnvStats function boxcox, which computes objective values for user-specified powers, or computes the optimal power for the specified objective.

Details

Objects of class "boxcox" are lists that contain information about the powers that were used, the objective that was used, the values of the objective for the given powers, and whether an optimization was specified.

Methods

Generic functions that have methods for objects of class "boxcox" include:
plot, print.

Value

Required Components
The following components must be included in a legitimate list of class "boxcox".

lambda

Numeric vector containing the powers used in the Box-Cox transformations. If the value of the optimize component is FALSE, then lambda contains the values of all of the powers at which the objective was evaluated. If the value of the optimize component is TRUE, then lambda is a scalar containing the value of the power that maximizes the objective.

objective

Numeric vector containing the value(s) of the objective for the given value(s) of \(\lambda\) that are stored in the component lambda.

objective.name

character string indicating the objective that was used. The possible values are "PPCC" (probability plot correlation coefficient; the default), "Shapiro-Wilk" (the Shapiro-Wilk goodness-of-fit statistic), and "Log-Likelihood" (the log-likelihood function).

optimize

logical scalar indicating whether the objective was simply evaluted at the given values of lambda (optimize=FALSE), or instead the optimal power transformation was computed within the bounds specified by lambda (optimize=TRUE).

optimize.bounds

Numeric vector of length 2 with a names attribute indicating the bounds within which the optimization took place. When optimize=FALSE, this contains missing values.

eps

finite, positive numeric scalar indicating what value of eps was used. When the absolute value of lambda is less than eps, lambda is assumed to be 0 for the Box-Cox transformation.

sample.size

Numeric scalar indicating the number of finite, non-missing observations.

data.name

The name of the data object used for the Box-Cox computations.

bad.obs

The number of missing (NA), undefined (NaN) and/or infinite (Inf, -Inf) values that were removed from the data object prior to performing the Box-Cox computations.

Optional Component
The following component may optionally be included in a legitimate list of class "boxcox". It must be included if you want to call the function plot.boxcox and specify Q-Q plots or Tukey Mean-Difference Q-Q plots.

data

Numeric vector containing the data actually used for the Box-Cox computations (i.e., the original data without any missing or infinite values).

Note

Since objects of class "boxcox" are lists, you may extract their components with the $ and [[ operators.

Author

Steven P. Millard (EnvStats@ProbStatInfo.com)

See also

Examples

  # Create an object of class "boxcox", then print it out.
  # (Note: the call to set.seed simply allows you to reproduce this example.)

  set.seed(250) 
  x <- rlnormAlt(30, mean = 10, cv = 2) 

  dev.new()
  hist(x, col = "cyan")

  boxcox.list <- boxcox(x)

  data.class(boxcox.list)
#> [1] "boxcox"
  #[1] "boxcox"
  
  names(boxcox.list)
#>  [1] "lambda"          "objective"       "objective.name"  "optimize"       
#>  [5] "optimize.bounds" "eps"             "data"            "sample.size"    
#>  [9] "data.name"       "bad.obs"        
  # [1] "lambda"          "objective"       "objective.name" 
  # [4] "optimize"        "optimize.bounds" "eps"            
  # [7] "data"            "sample.size"     "data.name"      
  #[10] "bad.obs" 

  boxcox.list
#> $lambda
#> [1] -2.0 -1.5 -1.0 -0.5  0.0  0.5  1.0  1.5  2.0
#> 
#> $objective
#> [1] 0.5423739 0.6402782 0.7818160 0.9272219 0.9921702 0.9581178 0.8749611
#> [8] 0.7827009 0.7004547
#> 
#> $objective.name
#> [1] "PPCC"
#> 
#> $optimize
#> [1] FALSE
#> 
#> $optimize.bounds
#> lower upper 
#>    NA    NA 
#> 
#> $eps
#> [1] 2.220446e-16
#> 
#> $data
#>  [1]  2.0192184  1.3268068 13.0050656 14.6964419  4.4758993  2.8075661
#>  [7]  2.7475058  4.5005255  2.2994106  3.6594709  9.2627400  8.0372560
#> [13]  3.6735836  3.7105992  3.4081524 13.4421468  0.9784090  2.1583361
#> [19]  4.1931244  6.3832010  9.2376357  3.2546640  2.6481660 12.4014044
#> [25]  0.4022087 22.4564613  1.0683931 30.0176128  8.8987899  1.5006026
#> 
#> $sample.size
#> [1] 30
#> 
#> $data.name
#> [1] "x"
#> 
#> $bad.obs
#> [1] 0
#> 
#> attr(,"class")
#> [1] "boxcox"
  #Results of Box-Cox Transformation
  #---------------------------------
  #
  #Objective Name:                  PPCC
  #
  #Data:                            x
  #
  #Sample Size:                     30
  #
  # lambda      PPCC
  #   -2.0 0.5423739
  #   -1.5 0.6402782
  #   -1.0 0.7818160
  #   -0.5 0.9272219
  #    0.0 0.9921702
  #    0.5 0.9581178
  #    1.0 0.8749611
  #    1.5 0.7827009
  #    2.0 0.7004547

  boxcox(x, optimize = TRUE) 
#> $lambda
#> [1] 0.04530789
#> 
#> $objective
#> [1] 0.9925919
#> 
#> $objective.name
#> [1] "PPCC"
#> 
#> $optimize
#> [1] TRUE
#> 
#> $optimize.bounds
#> lower upper 
#>    -2     2 
#> 
#> $eps
#> [1] 2.220446e-16
#> 
#> $data
#>  [1]  2.0192184  1.3268068 13.0050656 14.6964419  4.4758993  2.8075661
#>  [7]  2.7475058  4.5005255  2.2994106  3.6594709  9.2627400  8.0372560
#> [13]  3.6735836  3.7105992  3.4081524 13.4421468  0.9784090  2.1583361
#> [19]  4.1931244  6.3832010  9.2376357  3.2546640  2.6481660 12.4014044
#> [25]  0.4022087 22.4564613  1.0683931 30.0176128  8.8987899  1.5006026
#> 
#> $sample.size
#> [1] 30
#> 
#> $data.name
#> [1] "x"
#> 
#> $bad.obs
#> [1] 0
#> 
#> attr(,"class")
#> [1] "boxcox"
  #Results of Box-Cox Transformation
  #---------------------------------
  #
  #Objective Name:                  PPCC
  #
  #Data:                            x
  #
  #Sample Size:                     30
  #
  #Bounds for Optimization:         lower = -2
  #                                 upper =  2
  #
  #Optimal Value:                   lambda = 0.04530789
  #
  #Value of Objective:              PPCC = 0.9925919 

  #----------

  # Clean up
  #---------
  rm(x, boxcox.list)