S3 Class "calibrate"

Objects of S3 class "calibrate" are returned by the EnvStats function calibrate, which fits a calibration line or curve based on linear regression.

Details

Objects of class "calibrate" are lists that inherit from class "lm" and include a component called x that stores the model matrix (the values of the predictor variables for the final calibration model).

Methods

Generic functions that have methods for objects of class "calibrate" include:
NONE AT PRESENT.

Value

See the help file for lm.

Required Components
Besides the usual components in the list returned by the function lm, the following components must be included in a legitimate list of class "calibrate".

x

the model matrix from the linear model fit.

Note

Since objects of class "calibrate" are lists, you may extract their components with the $ and [[ operators.

Author

Steven P. Millard (EnvStats@ProbStatInfo.com)

See also

Examples

  # Create an object of class "calibrate", then print it out.

  # The data frame EPA.97.cadmium.111.df contains calibration data for 
  # cadmium at mass 111 (ng/L) that appeared in Gibbons et al. (1997b) 
  # and were provided to them by the U.S. EPA. 

  calibrate.list <- calibrate(Cadmium ~ Spike, data = EPA.97.cadmium.111.df) 
 
  names(calibrate.list)
#>  [1] "coefficients"  "residuals"     "effects"       "rank"         
#>  [5] "fitted.values" "assign"        "qr"            "df.residual"  
#>  [9] "xlevels"       "call"          "terms"         "model"        
#> [13] "x"            

  calibrate.list
#> 
#> Call:
#> lm(formula = formula, data = data, x = TRUE)
#> 
#> Coefficients:
#> (Intercept)        Spike  
#>      1.6385       0.9731  
#> 

  #----------

  # Clean up
  #---------
  rm(calibrate.list)