Estimate Probability Parameter of a Negative Binomial Distribution

Estimate the probability parameter of a negative binomial distribution.

enbinom(x, size, method = "mle/mme")

Arguments

x

vector of non-negative integers indicating the number of trials that took place before size “successes” occurred. (The total number of trials that took place is x+1). Missing (NA), undefined (NaN), and infinite (Inf, -Inf) values are allowed but will be removed. If length(x)=n and n is greater than 1, it is assumed that x represents observations from n separate negative binomial experiments that all had the same probability of success (prob), but possibly different values of size.

size

vector of positive integers indicating the number of “successes” that must be observed before the trials are stopped. Missing (NA), undefined (NaN), and infinite (Inf, -Inf) values are allowed but will be removed. The length of size must be 1 or else the same length as x.

method

character string specifying the method of estimation. Possible values are:
"mle/mme" (maximum likelihood and method of moments; the default) and
"mvue" (minimum variance unbiased).
You cannot use method="mvue" if the sum of the elements in size is 1. See the DETAILS section for more information on these estimation methods.

Details

If x contains any missing (NA), undefined (NaN) or infinite (Inf, -Inf) values, they will be removed prior to performing the estimation.

Let \(\underline{x} = (x_1, x_2, \ldots, x_n)\) be a vector of \(n\) independent observations from negative binomial distributions with parameters prob=\(p\) and size=\(\underline{k}\), where where \(\underline{k} = c(k_1, k_2, \ldots, k_n)\) is a vector of \(n\) (possibly different) values.

It can be shown (e.g., Forbes et al., 2011) that if \(X\) is defined as: $$X = \sum^n_{i = 1} x_i$$ then \(X\) is an observation from a negative binomial distribution with parameters prob=\(p\) and size=\(K\), where $$K = \sum^n_{i = 1} k_i$$

Estimation
The maximum likelihood and method of moments estimator (mle/mme) of \(p\) is given by: $$\hat{p}_{mle} = \frac{K}{X + K}$$ and the minimum variance unbiased estimator (mvue) of \(p\) is given by: $$\hat{p}_{mvue} = \frac{K - 1}{X + K - 1}$$ (Forbes et al., 2011). Note that the mvue of \(p\) is not defined for \(K=1\).

Value

a list of class "estimate" containing the estimated parameters and other information.

See estimate.object for details.

References

Forbes, C., M. Evans, N. Hastings, and B. Peacock. (2011). Statistical Distributions. Fourth Edition. John Wiley and Sons, Hoboken, NJ.

Johnson, N. L., S. Kotz, and A. Kemp. (1992). Univariate Discrete Distributions. Second Edition. John Wiley and Sons, New York, Chapter 5.

Author

Steven P. Millard (EnvStats@ProbStatInfo.com)

Note

The negative binomial distribution has its roots in a gambling game where participants would bet on the number of tosses of a coin necessary to achieve a fixed number of heads. The negative binomial distribution has been applied in a wide variety of fields, including accident statistics, birth-and-death processes, and modeling spatial distributions of biological organisms.

The geometric distribution with parameter prob=\(p\) is a special case of the negative binomial distribution with parameters size=1 and prob=\(p\).

See also

Examples

  # Generate an observation from a negative binomial distribution with 
  # parameters size=2 and prob=0.2, then estimate the parameter prob. 
  # Note: the call to set.seed simply allows you to reproduce this example. 
  # Also, the only parameter that is estimated is prob; the parameter 
  # size is supplied in the call to enbinom.  The parameter size is printed in 
  # order to show all of the parameters associated with the distribution.

  set.seed(250) 
  dat <- rnbinom(1, size = 2, prob = 0.2) 
  dat
#> [1] 5
  #[1] 5

  enbinom(dat, size = 2)
#> $distribution
#> [1] "Negative Binomial"
#> 
#> $sample.size
#> [1] 1
#> 
#> $parameters
#>      size      prob 
#> 2.0000000 0.2857143 
#> 
#> $n.param.est
#> [1] 1
#> 
#> $method
#> [1] "mle/mme for 'prob'"
#> 
#> $data.name
#> [1] "dat, 2"
#> 
#> $bad.obs
#> [1] 0
#> 
#> attr(,"class")
#> [1] "estimate"
  #Results of Distribution Parameter Estimation
  #--------------------------------------------
  #
  #Assumed Distribution:            Negative Binomial
  #
  #Estimated Parameter(s):          size = 2.0000000
  #                                 prob = 0.2857143
  #
  #Estimation Method:               mle/mme for 'prob'
  #
  #Data:                            dat, 2
  #
  #Sample Size:                     1

  #----------

  # Generate 3 observations from negative binomial distributions with 
  # parameters size=c(2,3,4) and prob=0.2, then estimate the parameter 
  # prob using the mvue. 
  # (Note: the call to set.seed simply allows you to reproduce this example.)

  size.vec <- 2:4 
  set.seed(250) 
  dat <- rnbinom(3, size = size.vec, prob = 0.2) 
  dat 
#> [1]  5 19 12
  #[1]  5 19 12 

  enbinom(dat, size = size.vec, method = "mvue") 
#> $distribution
#> [1] "Negative Binomial"
#> 
#> $sample.size
#> [1] 3
#> 
#> $parameters
#>      size      prob 
#> 9.0000000 0.1818182 
#> 
#> $n.param.est
#> [1] 1
#> 
#> $method
#> [1] "mvue for 'prob'"
#> 
#> $data.name
#> [1] "dat, size.vec"
#> 
#> $bad.obs
#> [1] 0
#> 
#> attr(,"class")
#> [1] "estimate"
  #Results of Distribution Parameter Estimation
  #--------------------------------------------
  #
  #Assumed Distribution:            Negative Binomial
  #
  #Estimated Parameter(s):          size = 9.0000000
  #                                 prob = 0.1818182
  #
  #Estimation Method:               mvue for 'prob'
  #
  #Data:                            dat, size.vec
  #
  #Sample Size:                     3

  #----------

  # Clean up
  #---------
  rm(dat)