Estimate Parameters of a Pareto Distribution

Estimate the location and shape parameters of a Pareto distribution.

epareto(x, method = "mle", plot.pos.con = 0.375)

Arguments

x

numeric vector of observations.

method

character string specifying the method of estimation. Possible values are "mle" (maximum likelihood; the default), and "lse" (least-squares). See the DETAILS section for more information on these estimation methods.

plot.pos.con

numeric scalar between 0 and 1 containing the value of the plotting position constant used to construct the values of the empirical cdf. The default value is plot.pos.con=0.375. This argument is used only when method="lse".

Details

If x contains any missing (NA), undefined (NaN) or infinite (Inf, -Inf) values, they will be removed prior to performing the estimation.

Let \(\underline{x} = (x_1, x_2, \ldots, x_n)\) be a vector of \(n\) observations from a Pareto distribution with parameters location=\(\eta\) and shape=\(\theta\).

Maximum Likelihood Estimatation (method="mle")
The maximum likelihood estimators (mle's) of \(\eta\) and \(\theta\) are given by (Evans et al., 1993; p.122; Johnson et al., 1994, p.581): $$\hat{\eta}_{mle} = x_{(1)} \;\;\;\; (1)$$ $$\hat{\theta}_{mle} = n [\sum_{i=1}^n log(\frac{x_i}{\hat{\eta}_{mle}}) ]^{-1} \;\;\;\; (2)$$ where \(x_(1)\) denotes the first order statistic (i.e., the minimum value).

Least-Squares Estimation (method="lse")
The least-squares estimators (lse's) of \(\eta\) and \(\theta\) are derived as follows. Let \(X\) denote a Pareto random variable with parameters location=\(\eta\) and shape=\(\theta\). It can be shown that $$log[1 - F(x)] = \theta log(\eta) - \theta log(x) \;\;\;\; (3)$$ where \(F\) denotes the cumulative distribution function of \(X\). Set $$y_i = log[1 - \hat{F}(x_i)] \;\;\;\; (4)$$ $$z_i = log(x_i) \;\;\;\; (5)$$ where \(\hat{F}(x)\) denotes the empirical cumulative distribution function evaluated at \(x\). The least-squares estimates of \(\eta\) and \(\theta\) are obtained by solving the regression equation $$y_i = \beta_{0} + \beta_{1} z_i \;\;\;\; (6)$$ and setting $$\hat{\theta}_{lse} = -\hat{\beta}_{1} \;\;\;\; (7)$$ $$\hat{\eta}_{lse} = exp(\frac{\hat{\beta}_0}{\hat{\theta}_{lse}}) \;\;\;\; (8)$$ (Johnson et al., 1994, p.580).

Value

a list of class "estimate" containing the estimated parameters and other information.

See estimate.object for details.

References

Forbes, C., M. Evans, N. Hastings, and B. Peacock. (2011). Statistical Distributions. Fourth Edition. John Wiley and Sons, Hoboken, NJ.

Johnson, N. L., S. Kotz, and N. Balakrishnan. (1994). Continuous Univariate Distributions, Volume 1. Second Edition. John Wiley and Sons, New York.

Author

Steven P. Millard (EnvStats@ProbStatInfo.com)

Note

The Pareto distribution is named after Vilfredo Pareto (1848-1923), a professor of economics. It is derived from Pareto's law, which states that the number of persons \(N\) having income \(\ge x\) is given by: $$N = A x^{-\theta}$$ where \(\theta\) denotes Pareto's constant and is the shape parameter for the probability distribution.

The Pareto distribution takes values on the positive real line. All values must be larger than the “location” parameter \(\eta\), which is really a threshold parameter. There are three kinds of Pareto distributions. The one described here is the Pareto distribution of the first kind. Stable Pareto distributions have \(0 < \theta < 2\). Note that the \(r\)'th moment only exists if \(r < \theta\).

The Pareto distribution is related to the exponential distribution and logistic distribution as follows. Let \(X\) denote a Pareto random variable with location=\(\eta\) and shape=\(\theta\). Then \(log(X/\eta)\) has an exponential distribution with parameter rate=\(\theta\), and \(-log\{ [(X/\eta)^\theta] - 1 \}\) has a logistic distribution with parameters location=\(0\) and scale=\(1\).

The Pareto distribution has a very long right-hand tail. It is often applied in the study of socioeconomic data, including the distribution of income, firm size, population, and stock price fluctuations.

See also

Examples

  # Generate 30 observations from a Pareto distribution with parameters 
  # location=1 and shape=1 then estimate the parameters. 
  # (Note: the call to set.seed simply allows you to reproduce this example.)

  set.seed(250) 
  dat <- rpareto(30, location = 1, shape = 1) 
  epareto(dat) 
#> $distribution
#> [1] "Pareto"
#> 
#> $sample.size
#> [1] 30
#> 
#> $parameters
#> location    shape 
#> 1.009046 1.079850 
#> 
#> $method
#> [1] "mle"
#> 
#> $data.name
#> [1] "dat"
#> 
#> $bad.obs
#> [1] 0
#> 
#> attr(,"class")
#> [1] "estimate"

  #Results of Distribution Parameter Estimation
  #--------------------------------------------
  #
  #Assumed Distribution:            Pareto
  #
  #Estimated Parameter(s):          location = 1.009046
  #                                 shape    = 1.079850
  #
  #Estimation Method:               mle
  #
  #Data:                            dat
  #
  #Sample Size:                     30

  #----------

  # Compare the results of using the least-squares estimators:

  epareto(dat, method="lse")$parameters 
#> location    shape 
#> 1.085924 1.144180 
  #location    shape 
  #1.085924 1.144180

  #----------

  # Clean up
  #---------

  rm(dat)