S3 Class "gofOutlier"

Objects of S3 class "gofOutlier" are returned by the EnvStats function rosnerTest.

Details

Objects of S3 class "gofOutlier" are lists that contain information about the assumed distribution, the test statistics, the Type I error level, and the number of outliers detected.

Methods

Generic functions that have methods for objects of class "gofOutlier" include:
print.

Value

Required Components
The following components must be included in a legitimate list of class "gofOutlier".

distribution

a character string indicating the name of the assumed distribution (see
Distribution.df).

statistic

a numeric vector with a names attribute containing the names and values of the outlier test statistic for each outlier tested.

sample.size

a numeric scalar containing the number of non-missing observations in the sample used for the outlier test.

parameters

numeric vector with a names attribute containing the name(s) and value(s) of the parameter(s) associated with the test statistic given in the statistic component.

alpha

numeric scalar indicating the Type I error level.

crit.value

numeric vector containing the critical values associated with the test for each outlier.

alternative

character string indicating the alternative hypothesis.

method

character string indicating the name of the outlier test.

data

numeric vector containing the data actually used for the outlier test (i.e., the original data without any missing or infinite values).

data.name

character string indicating the name of the data object used for the goodness-of-fit test.

all.stats

data frame containing all of the results of the test.

Optional Components
The following component is included when the data object contains missing (NA), undefined (NaN) and/or infinite (Inf, -Inf) values.

bad.obs

numeric scalar indicating the number of missing (NA), undefined (NaN) and/or infinite (Inf, -Inf) values that were removed from the data object prior to performing the test for outliers.

Note

Since objects of class "gofOutlier" are lists, you may extract their components with the $ and [[ operators.

Author

Steven P. Millard (EnvStats@ProbStatInfo.com)

See also

Examples

  # Create an object of class "gofOutlier", then print it out. 
  # (Note: the call to set.seed simply allows you to reproduce 
  # this example.)

  set.seed(250) 

  dat <- c(rnorm(30, mean = 3, sd = 2), rnorm(3, mean = 10, sd = 1)) 

  gofOutlier.obj <- rosnerTest(dat, k = 4) 

  mode(gofOutlier.obj) 
#> [1] "list"
  #[1] "list" 

  class(gofOutlier.obj) 
#> [1] "gofOutlier"
  #[1] "gofOutlier" 

  names(gofOutlier.obj) 
#>  [1] "distribution" "statistic"    "sample.size"  "parameters"   "alpha"       
#>  [6] "crit.value"   "n.outliers"   "alternative"  "method"       "data"        
#> [11] "data.name"    "bad.obs"      "all.stats"   
  # [1] "distribution" "statistic"    "sample.size"  "parameters"  
  # [5] "alpha"        "crit.value"   "n.outliers"   "alternative" 
  # [9] "method"       "data"         "data.name"    "bad.obs"     
  #[13] "all.stats"

  gofOutlier.obj 
#> $distribution
#> [1] "Normal"
#> 
#> $statistic
#>      R.1      R.2      R.3      R.4 
#> 2.848514 3.086875 3.033044 2.380235 
#> 
#> $sample.size
#> [1] 33
#> 
#> $parameters
#> k 
#> 4 
#> 
#> $alpha
#> [1] 0.05
#> 
#> $crit.value
#> lambda.1 lambda.2 lambda.3 lambda.4 
#> 2.951949 2.938048 2.923571 2.908473 
#> 
#> $n.outliers
#> [1] 3
#> 
#> $alternative
#> [1] "Up to 4 observations are not\n                                 from the same Distribution."
#> 
#> $method
#> [1] "Rosner's Test for Outliers"
#> 
#> $data
#>  [1]  1.7464403  1.0844139  4.6828666  4.8756193  3.0013261  2.2660652
#>  [7]  2.2319743  3.0099761  1.9512939  2.6838376  4.1479000  3.9241760
#> [13]  2.6899056  2.7057111  2.5716726  4.7349795  0.6042089  1.8514776
#> [19]  2.8984421  3.5609229  4.1436215  2.4990257  2.1739180  4.6079370
#> [25] -0.7972275  5.5440110  0.7429143  6.0015194  4.0847068  1.2784673
#> [31] 10.1460427  8.7340527 10.7593656
#> 
#> $data.name
#> [1] "dat"
#> 
#> $bad.obs
#> [1] 0
#> 
#> $all.stats
#>   i   Mean.i     SD.i      Value Obs.Num    R.i+1 lambda.i+1 Outlier
#> 1 0 3.549744 2.531011 10.7593656      33 2.848514   2.951949    TRUE
#> 2 1 3.324444 2.209872 10.1460427      31 3.086875   2.938048    TRUE
#> 3 2 3.104392 1.856109  8.7340527      32 3.033044   2.923571    TRUE
#> 4 3 2.916737 1.560335 -0.7972275      25 2.380235   2.908473   FALSE
#> 
#> attr(,"class")
#> [1] "gofOutlier"

  #Results of Outlier Test
  #-------------------------
  #
  #Test Method:                     Rosner's Test for Outliers
  #
  #Hypothesized Distribution:       Normal
  #
  #Data:                            dat
  #
  #Sample Size:                     33
  #
  #Test Statistics:                 R.1 = 2.848514
  #                                 R.2 = 3.086875
  #                                 R.3 = 3.033044
  #                                 R.4 = 2.380235
  #
  #Test Statistic Parameter:        k = 4
  #
  #Alternative Hypothesis:          Up to 4 observations are not
  #                                 from the same Distribution.
  #
  #Type I Error:                    5%
  #
  #Number of Outliers Detected:     3
  #
  #  i   Mean.i     SD.i      Value Obs.Num    R.i+1 lambda.i+1 Outlier
  #1 0 3.549744 2.531011 10.7593656      33 2.848514   2.951949    TRUE
  #2 1 3.324444 2.209872 10.1460427      31 3.086875   2.938048    TRUE
  #3 2 3.104392 1.856109  8.7340527      32 3.033044   2.923571    TRUE
  #4 3 2.916737 1.560335 -0.7972275      25 2.380235   2.908473   FALSE

  #==========

  # Extract the data frame with all the test results
  #-------------------------------------------------

  gofOutlier.obj$all.stats
#>   i   Mean.i     SD.i      Value Obs.Num    R.i+1 lambda.i+1 Outlier
#> 1 0 3.549744 2.531011 10.7593656      33 2.848514   2.951949    TRUE
#> 2 1 3.324444 2.209872 10.1460427      31 3.086875   2.938048    TRUE
#> 3 2 3.104392 1.856109  8.7340527      32 3.033044   2.923571    TRUE
#> 4 3 2.916737 1.560335 -0.7972275      25 2.380235   2.908473   FALSE
  #  i   Mean.i     SD.i      Value Obs.Num    R.i+1 lambda.i+1 Outlier
  #1 0 3.549744 2.531011 10.7593656      33 2.848514   2.951949    TRUE
  #2 1 3.324444 2.209872 10.1460427      31 3.086875   2.938048    TRUE
  #3 2 3.104392 1.856109  8.7340527      32 3.033044   2.923571    TRUE
  #4 3 2.916737 1.560335 -0.7972275      25 2.380235   2.908473   FALSE

  #==========

  # Clean up
  #---------
  rm(dat, gofOutlier.obj)