Print Values

The EnvStats function print is a generic function for printing its argument and returning it invisibly (via invisible(x)). The function invokes particular methods which depend on the class of the first argument. The EnvStats function print.default simply calls the R functions print or print.default.

The EnvStats functions print and print.default have been created in order to comply with CRAN policies, because EnvStats contains a modified version of the R function print.htest.

print(x, ...)

# S3 method for default
print(x, ...)

Arguments

x

an object to be printed and used to select a printing method.

...

further arguments passed to or from other methods. See the help files for the R functions print and print.default for more information.

Details

The EnvStats function print.default simply calls the R functions print or print.default, depending on the class of the argument x.

The EnvStats functions print and print.default have been created in order to comply with CRAN policies, because EnvStats contains a modified version of the R function print.htest. When EnvStats is loaded, objects of class "htest" will be printed using the EnvStats version of print.htest.

See the help files for the R functions print and print.default.

Value

See the help files for the R functions print and print.default.

References

Chambers, J.M., and Hastie, T.J., eds. (1992). Statistical Models in S. Chapman and Hall/CRC, Boca Raton, FL.

Author

R Development Core Team for code for R versions of print and print.default.

Steven P. Millard for EnvStats version of print.default; EnvStats@ProbStatInfo.com)

See also